requirements.txt
openai
python-dotenv
beautifulsoup4
playwright
selenium